# perf: use force while importing polyfill to avoid calling shouldPolyfill which adds overhead