--- a/src/libs/IntlPolyfill/index.ios.ts
+++ b/src/libs/IntlPolyfill/index.ios.ts
@@ -11,11 +11,11 @@ const intlPolyfill: IntlPolyfill = () => {
     // not yet implemented in hermes.
     // see support: https://hermesengine.dev/docs/intl/
 
-    require('@formatjs/intl-locale/polyfill');
+    require('@formatjs/intl-locale/polyfill-force');
 
     // Required to polyfill NumberFormat on iOS
     // see: https://github.com/facebook/hermes/issues/1172#issuecomment-1776156538
-    require('@formatjs/intl-pluralrules/polyfill');
+    require('@formatjs/intl-pluralrules/polyfill-force');
     polyfillNumberFormat();
 
     // Required to polyfill DateTimeFormat on iOS