--- a/src/libs/IntlPolyfill/polyfillDateTimeFormat.ts
+++ b/src/libs/IntlPolyfill/polyfillDateTimeFormat.ts
@@ -13,7 +13,7 @@ export default function () {
         currentTimezone = timezoneBackwardMap[currentTimezone];
     }
 
-    require('@formatjs/intl-datetimeformat/polyfill');
+    require('@formatjs/intl-datetimeformat/polyfill-force');
     require('@formatjs/intl-datetimeformat/locale-data/en');
     require('@formatjs/intl-datetimeformat/locale-data/es');
     require('@formatjs/intl-datetimeformat/add-all-tz');