@Beamanator 2024-08-13
Thanks very much @hurali97 🙏 and sorry for the delay and I have not shared this in slack, please feel free 👍