@OSBotify 2024-08-13
🚀 [Deployed](https://github.com/Expensify/App/actions/runs/10376405594) to staging by https://github.com/Beamanator in version: 9.0.20-0 🚀

platform | result
---|---
🤖 android 🤖|success ✅
🖥 desktop 🖥|success ✅
🍎 iOS 🍎|success ✅
🕸 web 🕸|success ✅