@hurali97 2024-08-16
@Beamanator I am not sure about the test steps but I tested the PR to verify that the date is appearing correctly for report actions. Maybe we can do that 👍